updated examples for consistency/accessibility
Added labels to form elements and changed some <span> to <p> (to get rid of <br>). Changed the order of some of the examples so that the form inputs are consistently first and the reactive elements are second.

The true/false checkbox example is a semantic catch 22. I changed it a bit. It is not quite as concise, but I think it still gets the point across.